DF-12322 fix client compile
Former-commit-id: ce410b1e05c2f9f8392557afcf16c2a9091befe2

+++ Sources/Tools/ResourceArchiver/ResourceArchiver.cpp
@@ -603,7 +603,7 @@ bool Pack(const Vector<CollectedFile>& collectedFiles, DAVA::Compressor::Type co
     footerBlock.infoCrc32 = CRC32::ForBuffer(&footerBlock.info, sizeof(footerBlock.info));
     if (meta)
     {
-        footerBlock.sizeOfMetaData = metaBytes.size();
+        footerBlock.sizeOfMetaData = static_cast<uint32>(metaBytes.size());
     }
     else
     {
